requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlmodel>=0.0.8
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
python-multipart>=0.0.5
pydantic<2.0.0
python-dotenv>=0.19.0
